requirements.txt
torch>=2.1.0
torchaudio>=2.1.0
gradio>=4.0.0
numpy
transformers
tensorboard
librosa
torchfcpe
tqdm
faiss-cpu
torchcodec
matplotlib